+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 11/9/2023
+- Minor: Removed log if there's no exchange group
+
 ## 11/4/2023
 - Minor: Reworked new code checker and codes now check for hoyolab if there's any new codes too
 - Minor: Added option to auto-redeem for codes in `config.js` and added function to automatically redeem new codes